Extract new group route into file
